Extract shared Artifact types into shared/types.ts
Single source of truth for Artifact, ArtifactType, ArtifactStatus, and
IconStatus — imported by both server and web. Resolves type divergence:
- Server had type: string, client had a union — now both use ArtifactType
- Client was missing iconStatus field — now included
- Status values were duplicated — now defined once

Part of #6
